+++ tests/sources/bug_407.ml
@@ -2,15 +2,15 @@ open Elpi.API
 
 let global = Ast.Loc.initial "init"
 let base_str = 
-  {| kind foo type. 
-     external symbol mk_t : int -> foo = "1".
+  {| data foo.
+     builtin symb mk_t : int -> foo = "1".
 
-     type is_t foo -> int -> (pred).
+     pred is_t -> foo, int.
      is_t (mk_t N) N.
   |}
 
 let is_t_c = RawData.Constants.declare_global_symbol "is_t"
-let mk_t_c = RawData.Constants.declare_global_symbol "mk_t"
+let mk_t_c = RawData.Constants.declare_global_symbol ~variant:1 "mk_t"
 let query = 
   Ast.Term.mkAppGlobal ~loc:global ~hdloc:global is_t_c
     (Ast.Term.mkVar ~loc:global ~hdloc:global (Ast.Name.from_string "X") [])